Authentication and Authrization Using APIs

+++ controllers/api/v1/posts_api.js
@@ -1,7 +1,7 @@
 const Post = require('../../../models/post');
 const Comment = require('../../../models/comment');
 module.exports.index = async function(req, res) {
-  
+
   let posts = await Post.find({})
         .sort('-createdAt')
         .populate('user')
@@ -11,7 +11,7 @@ module.exports.index = async function(req, res) {
                 path: 'user'
             }
         });
-  
+
   return res.json(200, {
     message: "List of posts",
     posts: posts
@@ -24,7 +24,7 @@ module.exports.destroy = async function(req, res){
     try{
         let post = await Post.findById(req.params.id);
 
-        // if (post.user == req.user.id){
+        if (post.user == req.user.id){
             post.remove();
 
             await Comment.deleteMany({post: req.params.id});
@@ -32,7 +32,11 @@ module.exports.destroy = async function(req, res){
             return res.json(200, {
                 message: "Post and associated comments deleted successfully!"
             });
-
+        }else {
+            return res.json(401, {
+                message: 'You cannot delete this post!'
+            });
+        }
 
     }catch(err){
         console.log("*****", err);
@@ -40,5 +44,5 @@ module.exports.destroy = async function(req, res){
             message: "Internal Server Error"
         });
     }
-    
+
 }
